pyproject.toml
[project]
name = "medium"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "argon2-cffi>=25.1.0",
    "fastapi[standard]>=0.135.1",
    "jinja2>=3.1.6",
    "pydantic>=2.12.5",
    "pyjwt>=2.12.1",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "sqlmodel>=0.0.37",
]

[dependency-groups]
lint = [
    "djlint>=1.36.4",
    "ruff>=0.15.6",
]
